+++ scripts/install.sh
@@ -68,8 +68,9 @@ $mpremote fs cp -r builtin :/
 #$mpremote fs cp -r data/images :/data/
 
 $mpremote fs mkdir :/data
-$mpremote fs mkdir :/data/com.micropythonos.system.wifiservice
-$mpremote fs cp ../internal_filesystem_excluded/data/com.micropythonos.system.wifiservice/config.json :/data/com.micropythonos.system.wifiservice/
+$mpremote fs mkdir :/data/prefs
+$mpremote fs mkdir :/data/prefs/com.micropythonos.system.wifiservice
+$mpremote fs cp ../internal_filesystem_excluded/data/prefs/com.micropythonos.system.wifiservice/config.json :/data/prefs/com.micropythonos.system.wifiservice/
 
 $mpremote fs mkdir :/apps